Change processing support_confidence_interval property
